requirements.txt
pandas==2.1.4
requests==2.31.0
psycopg2-binary==2.9.9  <-- NEW: The SQL Driver